// integral/app/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Plan = {
  FREE: 'FREE',
  STARTER: 'STARTER',
  GROWTH: 'GROWTH',
  ENTERPRISE: 'ENTERPRISE'
} as const

export type Plan = (typeof Plan)[keyof typeof Plan]


export const MemberRole = {
  OWNER: 'OWNER',
  ADMIN: 'ADMIN',
  VIEWER: 'VIEWER'
} as const

export type MemberRole = (typeof MemberRole)[keyof typeof MemberRole]


export const IntegrationStatus = {
  ACTIVE: 'ACTIVE',
  EXPIRING_SOON: 'EXPIRING_SOON',
  EXPIRED: 'EXPIRED',
  ERROR: 'ERROR',
  DISCONNECTED: 'DISCONNECTED'
} as const

export type IntegrationStatus = (typeof IntegrationStatus)[keyof typeof IntegrationStatus]


export const IntegrationCategory = {
  CRM: 'CRM',
  COMMUNICATION: 'COMMUNICATION',
  PRODUCTIVITY: 'PRODUCTIVITY',
  DEVTOOLS: 'DEVTOOLS',
  ANALYTICS: 'ANALYTICS',
  FINANCE: 'FINANCE',
  HR: 'HR',
  SECURITY: 'SECURITY',
  OTHER: 'OTHER'
} as const

export type IntegrationCategory = (typeof IntegrationCategory)[keyof typeof IntegrationCategory]


export const DiscoverySource = {
  GOOGLE_WORKSPACE: 'GOOGLE_WORKSPACE',
  OKTA: 'OKTA',
  AZURE_AD: 'AZURE_AD',
  STRIPE_BILLING: 'STRIPE_BILLING',
  MANUAL: 'MANUAL'
} as const

export type DiscoverySource = (typeof DiscoverySource)[keyof typeof DiscoverySource]


export const AuthType = {
  OAUTH2: 'OAUTH2',
  API_KEY: 'API_KEY',
  SAML: 'SAML',
  BASIC: 'BASIC'
} as const

export type AuthType = (typeof AuthType)[keyof typeof AuthType]


export const JobStatus = {
  PENDING: 'PENDING',
  RUNNING: 'RUNNING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED'
} as const

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus]